+++ src/config/index.test.ts
@@ -1,4 +1,8 @@
-import { getWaybackServiceBaseURL, setDefaultWaybackOptions } from './';
+import {
+    getWaybackServiceBaseURL,
+    setDefaultWaybackOptions,
+    getTileImageURL,
+} from './';
 
 import {
     WAYBACK_SERVICE_URL_TEMPLATE,
@@ -7,10 +11,6 @@ import {
 } from './';
 
 describe('test getWaybackServiceBaseURL', () => {
-    afterEach(() => {
-        jest.clearAllMocks();
-    });
-
     test('should return a URL from production subdomains', () => {
         const result = getWaybackServiceBaseURL();
 
@@ -47,6 +47,49 @@ describe('test getWaybackServiceBaseURL', () => {
 
         const expectedUrls = WAYBACK_SERVICE_SUB_DOMAINS_DEV.map((subDomain) =>
             WAYBACK_SERVICE_URL_TEMPLATE.replace('{subDomain}', subDomain)
+        );
+
+        expect(expectedUrls).toContain(result);
+    });
+});
+
+describe('test getTileImageURL', () => {
+    test('should return a URL from production subdomains with correct column, row and level', () => {
+        const result = getTileImageURL({
+            urlTemplate:
+                'https://wayback.maptiles.arcgis.com/arcgis/rest/services/World_Imagery/WMTS/1.0.0/default028mm/MapServer/tile/10/{level}/{row}/{col}',
+            column: 0,
+            row: 0,
+            level: 0,
+        });
+
+        const RESULT_TEMPLATE = `https://wayback.maptiles.arcgis.com/arcgis/rest/services/World_Imagery/WMTS/1.0.0/default028mm/MapServer/tile/10/0/0/0`;
+
+        const expectedUrls = WAYBACK_SERVICE_SUB_DOMAINS_PROD.map((subDomain) =>
+            RESULT_TEMPLATE.replace('wayback', subDomain)
+        );
+
+        expect(expectedUrls).toContain(result);
+    });
+
+    test('should return a URL from development subdomains with correct column, row and level', () => {
+        // Mock shouldUseDevServices to return true
+        setDefaultWaybackOptions({
+            useDevServices: true,
+        });
+
+        const result = getTileImageURL({
+            urlTemplate:
+                'https://waybackdev.maptiles.arcgis.com/arcgis/rest/services/World_Imagery/WMTS/1.0.0/default028mm/MapServer/tile/10/{level}/{row}/{col}',
+            column: 0,
+            row: 0,
+            level: 0,
+        });
+
+        const RESULT_TEMPLATE = `https://waybackdev.maptiles.arcgis.com/arcgis/rest/services/World_Imagery/WMTS/1.0.0/default028mm/MapServer/tile/10/0/0/0`;
+
+        const expectedUrls = WAYBACK_SERVICE_SUB_DOMAINS_DEV.map((subDomain) =>
+            RESULT_TEMPLATE.replace('waybackdev', subDomain)
         );
 
         expect(expectedUrls).toContain(result);