+++ CHANGELOG.md
@@ -13,5 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - update `getWaybackServiceBaseURL` to return URLs with different sub domains.
+- update `getTileImageURL` to return URLs with different sub domains. 
 
 ### Removed